--- lib/react_native_util/rake_task.rb
@@ -1,22 +1,0 @@
-require 'rake'
-require 'rake/tasklib'
-require_relative 'converter'
-
-module ReactNativeUtil
-  class RakeTask < Rake::TaskLib
-    include Util
-    def initialize(
-      name = :react_pod,
-      description = 'Convert project to use React pod',
-      chdir: '.',
-      repo_update: boolean_env_var?(:REACT_NATIVE_UTIL_REPO_UPDATE)
-    )
-      desc description
-      task name do
-        Dir.chdir chdir do
-          Converter.new(repo_update: repo_update).convert_to_react_pod!
-        end
-      end
-    end
-  end
-end
